+++ ChatterBox/try-app/src/container/ChatForm/chatFrom.js
@@ -1,6 +1,6 @@
 import SideMenu from '../side-menu/sidemenu.js'
 import Header from '../Header'
-import Settings from '../settings/settings.js'
+import Settings from '../profile-settings/profile-settings.js'
 import { push } from 'react-router-redux'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
