+++ ChatterBox/try-app/src/container/app/index.js
@@ -8,7 +8,7 @@ import StepLast from '../setProfile/stepLast.js'
 import chatForm from '../chat-form/chat-form.js'
 import SideMenu from '../side-menu/sidemenu.js'
 import Header from '../Header'
-import Settings from '../settings/settings.js'
+import Settings from '../profile-settings/profile-settings.js'
 import ChatForm from '../ChatForm/chatFrom.js'
 
 const App = () => (
